feat: implement bucket resource (and import) and datasource
For now, we also provide terraform files for testing in the
`/tests/terraform/bucket` folder.
Buckets to test resource import and datasource read must already be
created in the StorageGRID before running the tests.
Failures will be expected otherwise.
Name and region for read and imported buckets can be configured via
terraform variables to give testers some flexibility, particularly, in
terms of the buckets' regions.