Implement Bucket Versioning and Object Locking (#31)
* feat: implement bucket versioning resource (and import) and datasource

We also provide terraform files for testing in the
`/tests/terraform/bucket` folder.
As this resource further configures buckets, we append the versioning
resource tests to the regular bucket tests.
Therefore, buckets to test resource imports and datasource reads must
already be created in the StorageGRID before running the tests.
Failures will be expected otherwise.
Name and region for read and imported buckets can be configured via
terraform variables to give testers some flexibility, particularly, in
terms of the buckets' regions.
The bucket versioning resource relies on those names and does not need
to be configured additionally.

This commit also contains traces of linting and formatting in otherwise
completely unrelated files.

* feat: extend bucket resource with object locking

Object locking cannot be modified once a bucket is created, which is the
reason object locking is not implemented as a separate resource, but as
a block definition the `storagegrid_bucket` resource and data source.

All existing CRUD operations, tests, examples and documentation have
been updated to reflect the new capability of the `storagegrid_bucket`
resource and data source.

* refactor: create a dedicated api client for bucket

Particularly with respect to the read functionality of bucket
information, which is shared among the Terraform resource and data
source for buckets, this saves a lot of duplicated code.

To provide a coherent client implementation, we also add methods for
creating, updating and deleting buckets to the dedicated api client.

* docs: add more specific documentation for the bucket tests