Provider can manage object quota and policies on buckets (#34)
* feat: implement object quota on bucket

With the object quota, we follow the same implementation approach that
has been chosen for the bucket versioning.
In contrast, though, the delete operation sets the object quota for the
bucket to infinite, i.e., there is no quota anymore on the bucket.

relates to #25

* refactor: bucket versioning implementation details (code reuse)

We follow the approach of the bucket quota implementation, which yields
the advantage to also share implementation among data sources and
resources as well as within the resource implementation.

* feat: implement bucket policy

With the bucket policy, we stayed faithful to the general idea of
implementing "additional" bucket resources as individual Terraform
resources if the StorageGrid API permits.
It permits for the bucket policies.

We implement the full set of features that can be implemented using
policies on buckets.

It is currently not supported to specify the policy as JSON and use
the `jsondecode` function to specify the policy.
We foresee different issues with attribute parsing for various use
cases.