deps(tools): terraform-plugin-docs v0.24.0 (#35)
After upgrading the terraform-plugin-docs dependency (and related
indirect ones) running `go mod tidy` enforces go 1.25.0 in the
`mod`-file.

This is now safe to do as the `fedora:latest` image now has go 1.25.0 in
its packages.
Therefore, when building and running an image from the Dockerfile, the
tooling now works within the container.

Signed-off-by: Martin Weissbach <martin@cmdaltent.net>